+++ src/Body/RawPayload.php
@@ -19,9 +19,6 @@ final class RawPayload implements PayloadInterface
         $this->payload = $payload;
     }
 
-    /**
-     * Get the header content type value.
-     */
     public function contentType(): ?string
     {
         return null;
@@ -114,9 +111,6 @@ final class RawPayload implements PayloadInterface
         return $this->payload === '';
     }
 
-    /**
-     * Get the string representation of the payload.
-     */
     public function __toString()
     {
         return $this->payload;